Large commit that introduces ability to read in teh saved model state for continuation. Also tidied the paramaters and person classes so that there are now no satic member initialisation and the population is initialised one at a time. This then allows the parameters object to be used as a variable within the chyanged person constructor. This then addresses possible cause for memory leaks when using update

+++ src/parameters.h
@@ -77,7 +77,7 @@ public:
 	double g_PM;
 	double g_dCM;
 	double g_mean_maternal_immunity;
-	int g_sum_maternal_immunity;
+	double g_sum_maternal_immunity;
 	int g_total_mums;
 	// contributions to infectious reservoir by state and age;
 	double g_g1;
